Factor out static helper methods from Worker into WorkerUtil
